requirements-test.txt
pytest==5.4.3
pytest-django==3.9.0
pytest-cov==2.10.0
